pyproject.toml
[tool.poetry]
name = "stslab-dev"
version = "0.0.1"
description = "StsDev tool for the development of StackState Agent Integrations"
authors = ["Ravan Naidoo <rnaidoo@stackstate.com>"]

[tool.poetry.scripts]
stsdev = "stslab_dev.cli.main:main"

[tool.poetry.dependencies]
python = "^3.7"
typer = "^0.3.2"
plumbum = "^1.7.0"
python-dotenv = "^0.15.0"
requests = "^2.25.1"
stringcase = "^1.2.0"
flakehell = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = { version = "*", allow-prereleases = true }
typer-cli = "^0.0.11"

[tool.pytest.ini_options]
testpaths = [
    "tests",
    "integration",
]

[tool.flakehell]
exclude = ["README.rst", "README.md", ".eggs", ".tox", "build",".venv"]
include = ["src", "tests"]
format = "colored"
max_line_length = 120
show_source = true

[tool.flakehell.plugins]
flake8-bandit = ["+*", "-S322"]
flake8-bugbear = ["+*"]
flake8-builtins = ["+*"]
flake8-pytest-style = ["+*"]
mccabe = ["+*"]
pep8-naming = ["+*"]
pycodestyle = ["+*"]
pyflakes = ["+*"]
pylint = ["+*"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
